streamlit_chatbot/requirements.txt
google-generativeai #if using GEMINI API
streamlit #most often required
pandas #if dealing with data